Move back request and response structures
