+++ crates/subspace-networking/src/request_handlers.rs
@@ -1,2 +1,3 @@
 pub mod generic_request_handler;
 pub mod object_mappings;
+pub mod pieces_by_range;
